+++ test/testengine/test_willoughbyengine.py
@@ -1,0 +1,17 @@
+import unittest
+from engine.willoughby_engine import WilloughbyEngine
+
+class TestWilloughbyEngine(unittest.TestCase):
+    def test_engine_should_be_serviced(self):
+        current_mileage = 60001
+        last_service_mileage = 0
+
+        engine = WilloughbyEngine(current_mileage, last_service_mileage)
+        self.assertTrue(engine.needs_service())
+
+    def test_engine_should_not_be_serviced(self):
+        current_mileage = 60000
+        last_service_mileage = 0
+
+        engine = WilloughbyEngine(current_mileage, last_service_mileage)
+        self.assertFalse(engine.needs_service())
